+++ fluentui_icons/src/main/java/com/microsoft/fluentui/icons/appbaricons/__AppBarIcons.kt
@@ -1,18 +1,18 @@
 package com.microsoft.fluentui.icons.appbaricons
 
 import androidx.compose.ui.graphics.vector.ImageVector
-import com.microsoft.fluentui.icons.appbaricons.appbarnavigationicon.Appbarnavigationicon
+import com.microsoft.fluentui.icons.appbaricons.appbaricons.Arrowback
 import kotlin.collections.List as ____KtList
 
-public object AppBarNavigationIcon
+public object AppBarIcons
 
 private var __AllIcons: ____KtList<ImageVector>? = null
 
-public val AppBarNavigationIcon.AllIcons: ____KtList<ImageVector>
+public val AppBarIcons.AllIcons: ____KtList<ImageVector>
   get() {
     if (__AllIcons != null) {
       return __AllIcons!!
     }
-    __AllIcons= listOf(Appbarnavigationicon)
+    __AllIcons= listOf(Arrowback)
     return __AllIcons!!
   }
